# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: LawVise API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /lawvise/upload:
    post:
      operationId: uploadDocument
      tags: [lawvise]
      summary: Upload a document file (PDF, DOCX, image) and extract text via OCR
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadDocumentBody"
      responses:
        "200":
          description: Extracted text and storage path
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadDocumentResponse"
...
  /lawvise/documents:
    get:
      operationId: getDocuments
      tags: [lawvise]
      summary: List user's saved documents
      parameters:
        - name: folderId
          in: query
          schema:
            type: ["integer", "null"]
        - name: search
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List of documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LegalDocument"
    post:
      operationId: saveDocument
      tags: [lawvise]
      summary: Save a document to vault
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LegalDocumentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LegalDocument"
  /lawvise/documents/{id}:
    patch:
      operationId: updateDocument
      tags: [lawvise]
      summary: Update a document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LegalDocumentUpdate"
      responses:
        "200":
          description: Updated document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LegalDocument"
    delete:
      operationId: deleteDocument
      tags: [lawvise]
      summary: Delete a document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /lawvise/folders:
    get:
      operationId: getFolders
      tags: [lawvise]
      summary: List user's document folders
      responses:
        "200":
          description: List of folders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DocumentFolder"
    post:
      operationId: createFolder
      tags: [lawvise]
      summary: Create a new folder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFolderBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentFolder"
  /lawvise/folders/{id}:
    patch:
      operationId: updateFolder
      tags: [lawvise]
      summary: Rename a folder
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateFolderBody"
      responses:
        "200":
          description: Updated folder
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentFolder"
    delete:
      operationId: deleteFolder
      tags: [lawvise]
      summary: Delete a folder
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /lawvise/cases:
    get:
      operationId: getCases
      tags: [lawvise]
      summary: List user's cases
      responses:
        "200":
          description: List of cases
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LegalCase"
    post:
      operationId: createCase
      tags: [lawvise]
      summary: Create a new case
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LegalCaseInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LegalCase"
  /lawvise/cases/{id}:
    patch:
      operationId: updateCase
      tags: [lawvise]
      summary: Update a case
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LegalCaseUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LegalCase"
    delete:
      operationId: deleteCase
      tags: [lawvise]
      summary: Delete a case
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /storage/uploads/request-url:
    post:
      operationId: requestUploadUrl
      tags: [storage]
      summary: Request a presigned upload URL for file storage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequestUploadUrlBody"
      responses:
        "200":
          description: Presigned upload URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequestUploadUrlResponse"
components:
  schemas:
...
    UploadDocumentBody:
      type: object
      required: [fileBase64, mimeType, fileName]
      properties:
        fileBase64:
          type: string
          description: Base64-encoded file content
        mimeType:
          type: string
          description: "MIME type: application/pdf, application/vnd.openxmlformats-officedocument.wordprocessingml.document, image/jpeg, image/png, etc."
        fileName:
          type: string
        language:
          type: ["string", "null"]
    UploadDocumentResponse:
      type: object
      required: [extractedText, fileName, mimeType]
      properties:
        extractedText:
          type: string
        fileName:
          type: string
        mimeType:
          type: string
        pageCount:
          type: ["integer", "null"]
        objectPath:
          type: ["string", "null"]
...
    LegalDocument:
      type: object
      required: [id, title, documentType, createdAt, updatedAt]
      properties:
        id:
          type: integer
        userId:
          type: string
        title:
          type: string
        documentType:
          type: string
        content:
          type: ["string", "null"]
        analysisType:
          type: ["string", "null"]
        analysisResult:
          type: ["string", "null"]
        fileUrl:
          type: ["string", "null"]
        folderId:
          type: ["integer", "null"]
        isFavorite:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
    LegalDocumentInput:
      type: object
      required: [title, documentType]
      properties:
        title:
          type: string
        documentType:
          type: string
        content:
          type: ["string", "null"]
        analysisType:
          type: ["string", "null"]
        analysisResult:
          type: ["string", "null"]
        fileUrl:
          type: ["string", "null"]
        folderId:
          type: ["integer", "null"]
    LegalDocumentUpdate:
      type: object
      properties:
        title:
          type: string
        folderId:
          type: ["integer", "null"]
        isFavorite:
          type: boolean
        analysisResult:
          type: ["string", "null"]
    DocumentFolder:
      type: object
      required: [id, name, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: string
        name:
          type: string
        color:
          type: ["string", "null"]
        createdAt:
          type: string
    CreateFolderBody:
      type: object
      required: [name]
      properties:
        name:
          type: string
        color:
          type: ["string", "null"]
    UpdateFolderBody:
      type: object
      properties:
        name:
          type: string
        color:
          type: ["string", "null"]
    LegalCase:
      type: object
      required: [id, title, status, createdAt, updatedAt]
      properties:
        id:
          type: integer
        userId:
          type: string
        title:
          type: string
        caseNumber:
          type: ["string", "null"]
        court:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, pending, closed, won, lost]
        description:
          type: ["string", "null"]
        hearingDate:
          type: ["string", "null"]
        nextAction:
          type: ["string", "null"]
        priority:
          type: ["string", "null"]
          enum: [low, medium, high, null]
        createdAt:
          type: string
        updatedAt:
          type: string
    LegalCaseInput:
      type: object
      required: [title, status]
      properties:
        title:
          type: string
        caseNumber:
          type: ["string", "null"]
        court:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, pending, closed, won, lost]
        description:
          type: ["string", "null"]
        hearingDate:
          type: ["string", "null"]
        nextAction:
          type: ["string", "null"]
        priority:
          type: ["string", "null"]
          enum: [low, medium, high, null]
    LegalCaseUpdate:
      type: object
      properties:
        title:
          type: string
        caseNumber:
          type: ["string", "null"]
        court:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, pending, closed, won, lost]
        description:
          type: ["string", "null"]
        hearingDate:
          type: ["string", "null"]
        nextAction:
          type: ["string", "null"]
        priority:
          type: ["string", "null"]
          enum: [low, medium, high, null]
...
    RequestUploadUrlBody:
      type: object
      required: [name, size, contentType]
      properties:
        name:
          type: string
        size:
          type: integer
        contentType:
          type: string
    RequestUploadUrlResponse:
      type: object
      required: [uploadURL, objectPath]
      properties:
        uploadURL:
          type: string
        objectPath:
          type: string
        metadata:
          type: object
          properties:
            name:
              type: string
            size:
              type: integer
            contentType:
              type: string